FU-280 refactor: use notices list as common container
